fix toolbar downloaded project
