+++ app/layout/projects/show/header/toolbar/component.tsx
@@ -7,7 +7,7 @@ import { usePlausible } from 'next-plausible';
 
 import { useMe } from 'hooks/me';
 import {
-  useProject, useSaveProjectDownload, useExportId, useDownloadProject,
+  useProject, useSaveProjectDownload,
 } from 'hooks/projects';
 import { useScenarios } from 'hooks/scenarios';
 import { useToasts } from 'hooks/toast';
@@ -30,7 +30,6 @@ export const Toolbar: React.FC<ToolbarProps> = () => {
   const plausible = usePlausible();
 
   const { data: projectData } = useProject(pid);
-  const { data: exportId } = useExportId(pid);
   const { user } = useMe();
 
   const {
@@ -43,7 +42,6 @@ export const Toolbar: React.FC<ToolbarProps> = () => {
   });
 
   const projectDownloadMutation = useSaveProjectDownload({});
-  const downloadProject = useDownloadProject({});
 
   const scenarioIds = useMemo(() => {
     return scenariosData?.map((scenario) => scenario.id);
@@ -53,29 +51,7 @@ export const Toolbar: React.FC<ToolbarProps> = () => {
     projectDownloadMutation.mutate({ id: `${pid}`, data: { scenarioIds } }, {
 
       onSuccess: () => {
-        downloadProject.mutate({
-          id: `${pid}`,
-          exportId: `${exportId}`,
-        }, {
-          onSuccess: () => {
 
-          },
-          onError: () => {
-            addToast('download-error', (
-              <>
-                <h2 className="font-medium">Error!</h2>
-                <ul className="text-sm">
-                  `Project $
-                  {projectData?.name}
-                  {' '}
-                  not downloaded. Try again.`
-                </ul>
-              </>
-            ), {
-              level: 'error',
-            });
-          },
-        });
       },
       onError: ({ e }) => {
         console.error('error --->', e);
@@ -109,8 +85,6 @@ export const Toolbar: React.FC<ToolbarProps> = () => {
     scenarioIds,
     addToast,
     projectData?.name,
-    exportId,
-    downloadProject,
     plausible,
     user.email,
     user.id,
