+++ app/layout/projects/show/status/actions/done.tsx
@@ -4,12 +4,14 @@ import React, {
 
 import { useRouter } from 'next/router';
 
-import { useSaveProject } from 'hooks/projects';
+import { useDownloadProject, useExportId, useSaveProject } from 'hooks/projects';
 import { useToasts } from 'hooks/toast';
 
 export const useProjectActionsDone = () => {
   const { query } = useRouter();
   const { pid } = query;
+
+  const { addToast } = useToasts();
 
   const projectMutation = useSaveProject({
     requestConfig: {
@@ -17,7 +19,8 @@ export const useProjectActionsDone = () => {
     },
   });
 
-  const { addToast } = useToasts();
+  const downloadProject = useDownloadProject({});
+  const { data: exportId } = useExportId(pid);
 
   const onDone = useCallback((JOB_REF) => {
     projectMutation.mutate({
@@ -43,17 +46,38 @@ export const useProjectActionsDone = () => {
     });
   }, [pid, projectMutation, addToast]);
 
-  const onDownloadProject = useCallback((JOB_REF) => {
+  const onDownloadProjectDone = useCallback((JOB_REF) => {
     projectMutation.mutate({
       id: `${pid}`,
       data: {
         metadata: {
           cache: new Date().getTime(),
+          // cacheByUserId: { userId, cache: new Date().getTime() }
         },
       },
     }, {
       onSuccess: () => {
         JOB_REF.current = null;
+        downloadProject.mutate({
+          id: `${pid}`,
+          exportId: `${exportId}`,
+        }, {
+          onSuccess: () => {
+
+          },
+          onError: () => {
+            addToast('download-error', (
+              <>
+                <h2 className="font-medium">Error!</h2>
+                <ul className="text-sm">
+                  Project not downloaded. Try again.
+                </ul>
+              </>
+            ), {
+              level: 'error',
+            });
+          },
+        });
       },
       onError: () => {
         addToast('onDownloadProject', (
@@ -65,11 +89,11 @@ export const useProjectActionsDone = () => {
         });
       },
     });
-  }, [pid, projectMutation, addToast]);
+  }, [pid, projectMutation, addToast, downloadProject, exportId]);
 
   return {
     default: onDone,
     planningUnits: onDone,
-    export: onDownloadProject,
+    export: onDownloadProjectDone,
   };
 };
